requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
colorama==0.4.5
discord.py==2.0.1
frozenlist==1.3.1
httpcore==0.9.1
idna==3.3
loguru==0.6.0
lxml==4.9.1
multidict==4.7.6
requests==2.28.1
soupsieve==2.3.2.post1
typing_extensions==4.3.0
urllib3==1.26.11
win32-setctime==1.1.0
yarl==1.8.1
